Rename `replicon_channels` module into `channels`
