+++ src/client.rs
@@ -11,11 +11,11 @@ use bytes::Bytes;
 use varint_rs::VarintReader;
 
 use crate::core::{
+    channels::{ReplicationChannel, RepliconChannels},
     command_markers::{CommandMarkers, EntityMarkers},
     common_conditions::{client_connected, client_just_connected, client_just_disconnected},
     ctx::{DespawnCtx, RemoveCtx, WriteCtx},
     replication_fns::ReplicationFns,
-    replicon_channels::{ReplicationChannel, RepliconChannels},
     replicon_tick::RepliconTick,
     Replicated,
 };
