+++ src/server.rs
@@ -23,11 +23,11 @@ use bevy::{
 };
 
 use crate::core::{
+    channels::{ReplicationChannel, RepliconChannels},
     common_conditions::{server_just_stopped, server_running},
     ctx::SerializeCtx,
     replication_fns::ReplicationFns,
     replication_rules::ReplicationRules,
-    replicon_channels::{ReplicationChannel, RepliconChannels},
     replicon_tick::RepliconTick,
     ClientId,
 };
